requirements.txt
Django==4.0.2
djangorestframework==3.13.1
hashids==1.3.1
environs==9.5.0
validators==0.18.2